Add Socket mode adapter as new default Deprecate the RTM version of the adapter.
Fixes #84
